% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_connect.R
\name{fb_connect}
\alias{fb_connect}
\title{Returns the connection to the database}
\usage{
fb_connect(
  dbname = "flybase",
  host = "chado.flybase.org",
  port = 5432,
  user = "flybase",
  password = "flybase",
  dplyr = FALSE,
  ...
)
}
\arguments{
\item{dbname}{(character) database name. default: flybase}

\item{host}{(character) database name. default: chado.flybase.org}

\item{port}{(integer) database name. default: 5432}

\item{user}{(character) database name. default: flybase}

\item{password}{(character) database name. default: flybase}

\item{dplyr}{(logical) if \code{TRUE}, returns class \code{src_postgres}; 
otherwise, an object of class \code{PostgreSQLConnection}. 
Default: \code{FALSE}}

\item{...}{further args passed on to \code{\link[DBI]{dbConnect}}}
}
\description{
Returns the connection to the database
}
\details{
Make a connection with the read only postgreSQL version of 
the flybase database. For intense use, one should downlaod the database 
and change this function.
}
\examples{
\dontrun{
(conn <- fb_connect())

# using with dplyr
library("dplyr")
(src <- fb_connect(dplyr = TRUE))
orgs <- src \%>\% tbl("organism")
orgs \%>\% filter(genus == "Antheraea")

tbl(src, sql('SELECT * FROM phenotype WHERE "phenotype_id" = 8511'))
}
}
