% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_get_stockname.r
\name{fb_get_stockname}
\alias{fb_get_stockname}
\title{Returns the flybase stock name when the FBID number is provided, only 
one FBID number can be given at a time}
\usage{
fb_get_stockname(conn, fbst)
}
\arguments{
\item{conn}{postgresql connection object, from a call to 
\code{\link{fb_connect}}}

\item{fbst}{(character) flybase stock number}
}
\value{
single character string of the stock name
}
\description{
Returns the flybase stock name when the FBID number is provided, only 
one FBID number can be given at a time
}
\examples{
\dontrun{
conn <- fb_connect()
fb_get_stockname(conn, fbst = "FBst0000002")
}
}
