% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/fb_name_from_CG.r
=======
% Please edit documentation in R/FB_name_from_CG.r
>>>>>>> ropensci/master
\name{fb_name_from_cg}
\alias{fb_name_from_cg}
\title{Returns the gene name when the CG number is provided, only 
one CG number can be given at a time}
\usage{
fb_name_from_cg(conn, cg)
}
\arguments{
\item{conn}{postgresql connection object, from a call to 
\code{\link{fb_connect}}}

\item{cg}{(character) CG number}
}
\value{
single character string of the gene name
}
\description{
Returns the gene name when the CG number is provided, only 
one CG number can be given at a time
}
\examples{
\dontrun{
conn <- fb_connect()
fb_name_from_cg(conn, cg = "CG43860")
}
}
