% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_name_from_FBID.R
\name{fb_name_from_fbid}
\alias{fb_name_from_fbid}
\title{Returns the genetic element name when the FBID number is provided, only 
one number can be given at a time}
\usage{
fb_name_from_fbid(conn, fbid)
}
\arguments{
\item{conn}{postgresql connection object, from a call to 
\code{\link{fb_connect}}}

\item{fbid}{(character) fb id number}
}
\value{
single character string of the gene name
}
\description{
Returns the genetic element name when the FBID number is provided, only 
one number can be given at a time
}
\examples{
\dontrun{
conn <- fb_connect()
fb_name_from_fbid(conn, fbid = "FBtp0000001")
fb_name_from_fbid(conn, fbid = "FBti0038368")
fb_name_from_fbid(conn, fbid = "FBba0000025")
}
}
