% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_name_from_CG.r
\name{fb_namefromCG}
\alias{fb_namefromCG}
\title{Returns the gene name when the CG number is provided, only one CG number can be given at a time}
\usage{
fb_namefromCG(CGnumb)
}
\arguments{
\item{CGnumb}{CG number}
}
\description{
Returns the gene name when the CG number is provided, only one CG number can be given at a time
}
\examples{
\dontrun{  
fb_namefromCG("CG43860")
}
}

