% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rflybase-package.R
\docType{package}
\name{rflybase-package}
\alias{rflybase}
\alias{rflybase-package}
\title{Interface to Flybase}
\description{
rflybase: A programmatic interface to Flybase.org
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}

Julien Colomb
}

